Remove Direct3D.
I want to see if we can create a truly GL-only version. The source code
includes utilities for Direct3D. It could be that we could install
Direct3D on the development machine only, but then generate executables
that don't require Direct3D, but I prefer to avoid all references to
Direct3D to see how far we can go.